payment_bot/requirements.txt
aiogram==3.13.1
aiosqlite==0.20.0
SQLAlchemy==2.0.36
python-dotenv==1.0.1
aiofiles==24.1.0
aiohttp==3.11.10
